Using quantile for centering
- Centered inverse information content now calculated with quantile
- Default is of course 0.5 (mean)
- WARNING: Because Rust quantile calculation is not identical to e.g.
  R, values differ somewhat. prot-scriber behaviour may change a bit.

+++ src/default.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 lazy_static! {
 
     /// The score assigned to non informative words:
-    pub static ref NON_INFORMATIVE_WORD_SCORE : f32 = 0.000001;
+    pub static ref NON_INFORMATIVE_WORD_SCORE : f64 = 0.000001;
 
     /// The default Blacklist of regular expressions used to check for non-informative words
     /// in the description to be excluded from scoring. If ANY of these expression matches
